requirements.txt
matplotlib==3.8.0
numpy==2.2.4
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.2
tqdm==4.65.0
